Controller settings are added.

+++ MagicTetris/src/org/MagicTetris/util/KeySettings.java
@@ -10,12 +10,12 @@ import java.awt.event.KeyEvent;
  */
 public class KeySettings {
 	
-	private int KEY_ROTATE;
-	private int KEY_LEFT;
-	private int KEY_RIGHT;
-	private int KEY_DOWN;
-	private int KEY_CHANGE_ITEM;
-	private int KEY_USE_ITEM;
+	private float KEY_ROTATE;
+	private float KEY_LEFT;
+	private float KEY_RIGHT;
+	private float KEY_DOWN;
+	private float KEY_CHANGE_ITEM;
+	private float KEY_USE_ITEM;
 	
 	public KeySettings() {
 		KEY_ROTATE = -1;
@@ -56,51 +56,51 @@ public class KeySettings {
 		TWO
 	}
 
-	public int getKEY_ROTATE() {
+	public float getKEY_ROTATE() {
 		return KEY_ROTATE;
 	}
 
-	public void setKEY_ROTATE(int rotate) {
+	public void setKEY_ROTATE(float rotate) {
 		KEY_ROTATE = rotate;
 	}
 
-	public int getKEY_LEFT() {
+	public float getKEY_LEFT() {
 		return KEY_LEFT;
 	}
 
-	public void setKEY_LEFT(int left) {
+	public void setKEY_LEFT(float left) {
 		KEY_LEFT = left;
 	}
 
-	public int getKEY_RIGHT() {
+	public float getKEY_RIGHT() {
 		return KEY_RIGHT;
 	}
 
-	public void setKEY_RIGHT(int right) {
+	public void setKEY_RIGHT(float right) {
 		KEY_RIGHT = right;
 	}
 
-	public int getKEY_DOWN() {
+	public float getKEY_DOWN() {
 		return KEY_DOWN;
 	}
 
-	public void setKEY_DOWN(int down) {
+	public void setKEY_DOWN(float down) {
 		KEY_DOWN = down;
 	}
 
-	public int getKEY_CHANGE_ITEM() {
+	public float getKEY_CHANGE_ITEM() {
 		return KEY_CHANGE_ITEM;
 	}
 
-	public void setKEY_CHANGE_ITEM(int change_item) {
+	public void setKEY_CHANGE_ITEM(float change_item) {
 		KEY_CHANGE_ITEM = change_item;
 	}
 
-	public int getKEY_USE_ITEM() {
+	public float getKEY_USE_ITEM() {
 		return KEY_USE_ITEM;
 	}
 
-	public void setKEY_USE_ITEM(int use_item) {
+	public void setKEY_USE_ITEM(float use_item) {
 		KEY_USE_ITEM = use_item;
 	}
